+++ backend/package.json
@@ -29,6 +29,7 @@
     "mongodb": "^6.4.0",
     "mongoose": "^8.2.1",
     "nodemon": "^3.1.0",
+    "shared-types": "file:../shared-types",
     "swagger-jsdoc": "^6.2.8",
     "swagger-ui-express": "^5.0.0",
     "ts-jest": "^29.1.2",
